backend/requirements.txt
requests
beautifulsoup4
lxml
python-dateutil
google-genai
python-dotenv
tortoise-orm
fastapi
pydantic
uvicorn
asyncpg